Add missing helpers import
mv several methods to the appropriate class
fix bug in with_tags
